Create setup.py for app initialization
Separated initial configuration of youtube_dl and metadata
from split.py to a new file: setup.py

Work in progress https://github.com/crisbal/Album-Splitter/issues/17
